scripts/requirements-dev.txt
# Development dependencies for Supabase initialization script
psycopg2-binary>=2.9.6
python-dotenv>=0.21.0
pytest>=7.2.0
pytest-cov>=4.0.0
black>=22.12.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.0.0
